requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
asyncpg
psycopg2-binary
openai-agents
mem0ai
python-multipart
pydub
soundfile
ffmpeg-python==0.2.0
websockets
aiofiles
loguru
black
isort
flake8
google-genai
librosa



pydub==0.25.1
pyaudio==0.2.14
wave
audioop


pyaudio==0.2.13
websockets==12.0
twilio==8.10.0
python-dotenv==1.0.0
aiohttp==3.9.1 